% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationAnalysis_Step1.r
\name{contigextract}
\alias{contigextract}
\title{Extracting contig information}
\usage{
contigextract(contigRefxmap, chrom, startPos, endPos)
}
\arguments{
\item{contigRefxmap}{character. Contig xmap data.}

\item{chrom}{character. Chromosome of interest.}

\item{startPos}{integer.Start position.}

\item{endPos}{integer. End position.}
}
\value{
dataframe with data extracted from xmap.
}
\description{
Extracting contig information
}
\examples{
refcontigXmap <- system.file("extdata", "Cmapdir/output/contigs/exp_refineFinal1_sv/merged_smaps/ContigRef.xmap", package="methometR")
modxmap <- readingXmap(refcontigXmap)
extractedInfo<-contigextract(contigRefxmap = modxmap, chrom = 4, startPos = 100000, endPos = 200000)
}
