% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationAnalysis_Step1.r
\name{extracting_modifying_methLabels}
\alias{extracting_modifying_methLabels}
\title{Main function to call the extraction, modification and quantification process.}
\usage{
extracting_modifying_methLabels(
  cmaploc_mol,
  untar = FALSE,
  cntNick,
  cntMeth,
  selectedRegion = TRUE,
  chrom,
  startPos,
  endPos,
  outputFormat = c("individualText", "multipleText", "dataFrame"),
  path,
  logFilename
)
}
\arguments{
\item{cmaploc_mol}{character. path to the cmap folder.}

\item{untar}{boolean. untar option. True or False/}

\item{cntNick}{integer. Nick label value.}

\item{cntMeth}{integer. Meth label value.}

\item{selectedRegion}{boolean. if TRUE, selected region is analyzed and 
the user provides chromosome and start and end point. If FALSE, whole
genome analysed.}

\item{chrom}{character. Chromosome number if selectedRegion is TRUE}

\item{startPos}{integer. Start of the query if selectedRegion is TRUE}

\item{endPos}{integer. End of the query if selectedRegion is TRUE}

\item{outputFormat}{Character. Choices are individualText, multipleText and
dataframe.}

\item{path}{Character. Path of the folder where log files needs to be written 
to.}

\item{logFilename}{character. Log file name.}
}
\value{
dataframe or text files containing the modified files.
}
\description{
Main function to call the extraction, modification and quantification process.
}
\examples{
\dontrun{
molcmap <- system.file("extdata", "SampMolecule.cmap", package="methometR")
cmap <- system.file("extdata", "SampContig.cmap", package="methometR")
bed<-molMethylRefLoc(nickrefLoc,molcmap,xmapdata, contigID, cntNick , 
   cntMeth)
}
}
