% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationAnalysis_Step1.r
\name{modmolcmap}
\alias{modmolcmap}
\title{Identifies and documents the methylation and nick labels}
\usage{
modmolcmap(molcmap, xmapdata, cntNick, cntMeth)
}
\arguments{
\item{molcmap}{character. Path to Bionano molecule cmap.}

\item{xmapdata}{character. Path to Bionano molecule contig xmap.}

\item{cntNick}{integer. Nick label value.}

\item{cntMeth}{integer. Meth label value.}
}
\value{
Data Frame with the molecule cmap, with the exact site number 
for methylation and nick labels.
}
\description{
Identifies and documents the methylation and nick labels
}
\examples{
xmap <- system.file("extdata", "Cmapdir/output/contigs/exp_refineFinal1/merged_smaps/SampContigMolecule.xmap", package="methometR")
cmap <- system.file("extdata", "Cmapdir/output/contigs/exp_refineFinal1/merged_smaps/SampMolecule_q.cmap", package="methometR")
modcmap <- readingCmap(cmap)
modxmap <- readingXmap(xmap)
modMolcmap <- modmolcmap(molcmap = modcmap,
  xmapdata = modxmap,
  cntNick = 1,
  cntMeth=2)
}
