% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationAnalysis_Step1.r
\name{nickReference}
\alias{nickReference}
\title{Mapping nick label location from contig to reference}
\usage{
nickReference(
  refxmap,
  refcmap,
  contigcmap,
  contigID,
  returnMethod = c("Text", "dataFrame"),
  hmdir,
  chrom
)
}
\arguments{
\item{refxmap}{character. contig reference xmap data.}

\item{refcmap}{character. reference cmap data.}

\item{contigcmap}{character. contig cmap data.}

\item{contigID}{character. query contig ID.}

\item{returnMethod}{character. Output method. Text or dataframe.}

\item{hmdir}{integer. Home Directory.}

\item{chrom}{integer. chromosome number.}
}
\value{
dataframe containing the molecule file, with the reference position
}
\description{
Mapping nick label location from contig to reference
}
\examples{
refcontigXmap <- system.file("extdata", "Cmapdir/output/contigs/exp_refineFinal1_sv/merged_smaps/ContigRef.xmap", package="methometR")
refCmap <- system.file("extdata", "Cmapdir/output/contigs/exp_refineFinal1_sv/merged_smaps/hg19ref_r.cmap", package="methometR")
Contigqcmap <- system.file("extdata", "Cmapdir/output/contigs/exp_refineFinal1_sv/merged_smaps/SampContig_q.cmap", package="methometR")
refxmapdat <- readingXmap(refcontigXmap)
refcmapdat <- readingCmap(refCmap)
contigcmapdat <- readingCmap(Contigqcmap)
nickRef<-nickReference(refxmap = refxmapdat, refcmap = refcmapdat, 
    contigcmap = contigcmapdat, contigID = 6701, 
   returnMethod =c("dataFrame"),  chrom = 4)
}
