% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationAnalysis_Step1.r
\name{readingCmap}
\alias{readingCmap}
\title{Reads Bionano Cmap files}
\usage{
readingCmap(cmap)
}
\arguments{
\item{cmap}{character. Path to Bionano cmap file.}
}
\value{
Data Frame Contains the cmap file information.
}
\description{
Reads Bionano Cmap files
}
\examples{
cmap <- system.file("extdata",  "Cmapdir/output/contigs/exp_refineFinal1/merged_smaps/SampMolecule_q.cmap", package="methometR")
modcmap <- readingCmap(cmap)
}
