% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationAnalysis_Step1.r
\name{readingXmap}
\alias{readingXmap}
\title{Reads Bionano xmap files}
\usage{
readingXmap(xmap)
}
\arguments{
\item{xmap}{character. Path to Bionano xmap file.}
}
\value{
Data Frame Contains the xmap file information.
}
\description{
Reads Bionano xmap files
}
\examples{
xmap <- system.file("extdata", "Cmapdir/output/contigs/exp_refineFinal1/merged_smaps/SampContigMolecule.xmap", package="methometR")
modxmap <- readingXmap(xmap)
}
