% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationAnalysis_Step1.r
\name{untaringdirectoy}
\alias{untaringdirectoy}
\title{Untarring the directory if needed}
\usage{
untaringdirectoy(cmaploc_mol)
}
\arguments{
\item{cmaploc_mol}{character. Path to Bionano compressed file.}
}
\value{
uncompressed folder with all the files for analysis
}
\description{
Untarring the directory if needed
}
\examples{
cmaploc_mol <- system.file("extdata", package="methometR")
untarredFolder <-untaringdirectoy(cmaploc_mol)
}
